+++ app/Scraper/AbstractSiteExtractor.php
@@ -75,7 +75,10 @@ abstract class AbstractSiteExtractor implements SiteExtractorContract
      *
      * @return null|string
      */
-    abstract public function glass(): ?string;
+    public function glass(): ?string
+    {
+        return null;
+    }
 
     /**
      * Array containing cocktail ingredients
@@ -89,7 +92,10 @@ abstract class AbstractSiteExtractor implements SiteExtractorContract
      *
      * @return null|string
      */
-    abstract public function garnish(): ?string;
+    public function garnish(): ?string
+    {
+        return null;
+    }
 
     /**
      * Array containing image information
@@ -97,6 +103,16 @@ abstract class AbstractSiteExtractor implements SiteExtractorContract
      * @return null|array{"url": string|null, "copyright": string|null}
      */
     abstract public function image(): ?array;
+
+    /**
+     * Cocktail method (shake, stir...)
+     *
+     * @return null|string
+     */
+    public function method(): ?string
+    {
+        return null;
+    }
 
     /**
      * Cocktail information as array
@@ -113,6 +129,7 @@ abstract class AbstractSiteExtractor implements SiteExtractorContract
             'instructions' => $this->instructions(),
             'garnish' => $this->garnish(),
             'tags' => $this->tags(),
+            'method' => $this->method(),
             'images' => [
                 $this->image()
             ],