+++ app/Scraper/Manager.php
@@ -6,8 +6,7 @@ namespace Kami\Cocktail\Scraper;
 
 use Throwable;
 use Kami\Cocktail\Exceptions\ScrapeException;
-use Kami\Cocktail\Scraper\Sites\SchemaScraper;
-use Kami\Cocktail\Scraper\Sites\MicrodataScraper;
+use Kami\Cocktail\Scraper\Sites\DefaultScraper;
 
 final class Manager
 {
@@ -18,6 +17,10 @@ final class Manager
         \Kami\Cocktail\Scraper\Sites\EricsCocktailGuide::class,
         \Kami\Cocktail\Scraper\Sites\HausAlpenz::class,
         \Kami\Cocktail\Scraper\Sites\TheDrinkBlog::class,
+        \Kami\Cocktail\Scraper\Sites\DiffordsGuide::class,
+        \Kami\Cocktail\Scraper\Sites\TheCocktailDB::class,
+        \Kami\Cocktail\Scraper\Sites\CocktailParty::class,
+        \Kami\Cocktail\Scraper\Sites\LiberAndCo::class,
     ];
 
     public function __construct(private readonly string $url)
@@ -34,15 +37,9 @@ final class Manager
             }
         }
 
-        // Fallback to microdata
-        try {
-            return resolve(MicrodataScraper::class, ['url' => $this->url]);
-        } catch (Throwable $e) {
-        }
-
         // Fallback to schema scraper
         try {
-            return resolve(SchemaScraper::class, ['url' => $this->url]);
+            return resolve(DefaultScraper::class, ['url' => $this->url]);
         } catch (Throwable $e) {
         }